+++ src/API/private-api.js
@@ -112,7 +112,7 @@ export default class PrivateAPI {
 		const docData = foundry.utils.deepClone(data);
 		const sourceActor = game.actors.get(doc.actorId);
 		let itemPileConfig = foundry.utils.mergeObject(
-			foundry.utils.deepClone(CONSTANTS.PILE_DEFAULTS),
+			foundry.utils.deepClone(PileUtilities.getPileDefaults()),
 			foundry.utils.getProperty(docData, CONSTANTS.FLAGS.PILE) ?? {}
 		)
 		if (doc.isLinked || !foundry.utils.hasProperty(docData, CONSTANTS.FLAGS.PILE)) {
@@ -1070,7 +1070,7 @@ export default class PrivateAPI {
 		const { itemDeltas, attributeDeltas } = await targetTransaction.commit();
 
 		if (targetItemPileFlags) {
-			const flags = PileUtilities.cleanFlagData(foundry.utils.mergeObject(CONSTANTS.PILE_DEFAULTS, targetItemPileFlags));
+			const flags = PileUtilities.cleanFlagData(foundry.utils.mergeObject(PileUtilities.getPileDefaults(), targetItemPileFlags));
 			await PileUtilities.updateItemPileData(targetActor, flags);
 		}
 
@@ -1270,7 +1270,7 @@ export default class PrivateAPI {
 
 				Helpers.custom_notify("A Default Item Pile has been added to your Actors list. You can configure the default look and behavior on it, or duplicate it to create different styles.")
 
-				let pileDataDefaults = foundry.utils.deepClone(CONSTANTS.PILE_DEFAULTS);
+				let pileDataDefaults = PileUtilities.getPileDefaults();
 
 				pileDataDefaults.enabled = true;
 				if (foundry.utils.isEmpty(itemPileFlags)) {
@@ -1916,7 +1916,7 @@ export default class PrivateAPI {
 		const targetActor = Utilities.getActor(dropData.target);
 		if (sourceActor && targetActor && sourceActor === targetActor) return;
 
-		if (dropData.target && PileUtilities.isItemPileMerchant(dropData.target)) return;
+		if (dropData.target && PileUtilities.isItemPileMerchant(dropData.target) && !dropData.target.isOwner) return;
 
 		const validItem = await PileUtilities.checkItemType(dropData.target, dropData.itemData.item);
 		if (!validItem) return;
@@ -2579,7 +2579,7 @@ export default class PrivateAPI {
 		userId = false,
 	} = {}) {
 
-		let items = await PileUtilities.rollTable({
+		let { items, currencies } = await PileUtilities.rollTable({
 			tableUuid: table,
 			formula: timesToRoll,
 			normalize: normalizeTable,
@@ -2595,6 +2595,10 @@ export default class PrivateAPI {
 				return Utilities.setItemQuantity(actualItem, item.quantity);
 			});
 			items = await this._addItems(targetActor, itemsToAdd, userId, { removeExistingActorItems });
+
+			if (currencies.length > 0) {
+				await this._addCurrencies(targetActor, PileUtilities.getStringFromCurrencies(currencies), userId)
+			}
 		}
 
 		return items;